MH-12885 Capture Agent Access Management
This patch adds the possibility to configure which user (groups) have
access to which capture agents. Whenever a new capture agent is
registered, a new role is generated which a user needs to have when
changing scheduling information affecting the capture agent.

This includes:

- Editing schedule information of scheduled events
- Changing the workflow settings of scheduled events
- Scheduling new events
- Deleting scheduled events
- Deleting capture agents

The administrator as well as the organization administrator always has
access to all capture agents. For other users, access can be configured
by adding/removing the generated roles using the admin UI.

+++ modules/admin-ui/src/main/java/org/opencastproject/adminui/endpoint/AbstractEventEndpoint.java
@@ -120,6 +120,7 @@ import org.opencastproject.security.api.UnauthorizedException;
 import org.opencastproject.security.api.User;
 import org.opencastproject.security.urlsigning.exception.UrlSigningException;
 import org.opencastproject.security.urlsigning.service.UrlSigningService;
+import org.opencastproject.security.util.SecurityUtil;
 import org.opencastproject.systems.OpencastConstants;
 import org.opencastproject.util.DateTimeSupport;
 import org.opencastproject.util.Jsons.Val;
@@ -337,8 +338,9 @@ public abstract class AbstractEventEndpoint {
                   @RestResponse(responseCode = SC_OK, description = "The event has been deleted."),
                   @RestResponse(responseCode = HttpServletResponse.SC_NOT_FOUND, description = "The event could not be found."),
                   @RestResponse(responseCode = HttpServletResponse.SC_UNAUTHORIZED, description = "If the current user is not authorized to perform this action") })
-  public Response deleteEvent(@PathParam("eventId") String id) throws NotFoundException, UnauthorizedException {
+  public Response deleteEvent(@PathParam("eventId") String id) throws UnauthorizedException, SearchIndexException {
     try {
+      checkAgentAccess(id);
       if (!getIndexService().removeEvent(id))
         return Response.serverError().build();
     } catch (NotFoundException e) {
@@ -362,7 +364,7 @@ public abstract class AbstractEventEndpoint {
           @RestResponse(description = "Events have been deleted", responseCode = HttpServletResponse.SC_OK),
           @RestResponse(description = "The list of ids could not be parsed into a json list.", responseCode = HttpServletResponse.SC_BAD_REQUEST),
           @RestResponse(description = "If the current user is not authorized to perform this action", responseCode = HttpServletResponse.SC_UNAUTHORIZED) })
-  public Response deleteEvents(String eventIdsContent) throws UnauthorizedException {
+  public Response deleteEvents(String eventIdsContent) throws UnauthorizedException, SearchIndexException {
     if (StringUtils.isBlank(eventIdsContent)) {
       return Response.status(Response.Status.BAD_REQUEST).build();
     }
@@ -384,6 +386,7 @@ public abstract class AbstractEventEndpoint {
     for (Object eventIdObject : eventIdsJsonArray) {
       String eventId = eventIdObject.toString();
       try {
+        checkAgentAccess(eventId);
         if (!getIndexService().removeEvent(eventId)) {
           result.addServerError(eventId);
         } else {
@@ -391,6 +394,8 @@ public abstract class AbstractEventEndpoint {
         }
       } catch (NotFoundException e) {
         result.addNotFound(eventId);
+      } catch (UnauthorizedException e) {
+        result.addUnauthorized(eventId);
       }
     }
     return Response.ok(result.toJson()).build();
@@ -542,6 +547,12 @@ public abstract class AbstractEventEndpoint {
       agentId = Opt.some(schedulingJson.getString(SCHEDULING_AGENT_ID_KEY));
       logger.trace("Updating agent id of event '{}' from '{}' to '{}'",
         event.getIdentifier(), technicalMetadata.getAgentId(), agentId);
+    }
+
+    // Check if we are allowed to re-schedule on this agent
+    SecurityUtil.checkAgentAccess(getSecurityService(), technicalMetadata.getAgentId());
+    if (agentId.isSome()) {
+      SecurityUtil.checkAgentAccess(getSecurityService(), agentId.get());
     }
 
     Opt<Date> start = Opt.none();
@@ -764,6 +775,7 @@ public abstract class AbstractEventEndpoint {
       } else {
         MediaPackage mediaPackage = getIndexService().getEventMediapackage(optEvent.get());
         mediaPackage = getAuthorizationService().setAcl(mediaPackage, AclScope.Episode, accessControlList).getA();
+        // We could check agent access here if we want to forbid updating ACLs for users without access.
         getSchedulerService().updateEvent(eventId, Opt.<Date> none(), Opt.<Date> none(), Opt.<String> none(),
                 Opt.<Set<String>> none(), some(mediaPackage), Opt.<Map<String, String>> none(),
                 Opt.<Map<String, String>> none(), Opt.<Opt<Boolean>> none(), SchedulerService.ORIGIN);
@@ -1464,6 +1476,8 @@ public abstract class AbstractEventEndpoint {
         if (caMetadataOpt.isNone() && workflowConfigOpt.isNone())
           return Response.noContent().build();
 
+        SecurityUtil.checkAgentAccess(getSecurityService(), optEvent.get().getAgentId());
+
         getSchedulerService().updateEvent(id, Opt.<Date> none(), Opt.<Date> none(), Opt.<String> none(),
                 Opt.<Set<String>> none(), Opt.<MediaPackage> none(), workflowConfigOpt, caMetadataOpt,
                 Opt.<Opt<Boolean>> none(), SchedulerService.ORIGIN);
@@ -1837,6 +1851,7 @@ public abstract class AbstractEventEndpoint {
   public Response updateEventOptOut(@PathParam("eventId") String eventId, @PathParam("optout") boolean optout)
           throws NotFoundException, UnauthorizedException {
     try {
+      checkAgentAccess(eventId);
       getIndexService().changeOptOutStatus(eventId, optout, getIndex());
       return Response.noContent().build();
     } catch (SchedulerException e) {
@@ -2109,7 +2124,7 @@ public abstract class AbstractEventEndpoint {
                   @RestResponse(responseCode = SC_BAD_REQUEST, description = "If the metadata is not set or couldn't be parsed") })
   public Response createNewEvent(@Context HttpServletRequest request) {
     try {
-      String result = getIndexService().createEvent(request);
+      String result = getIndexService().createEvent(request, true);
       return Response.status(Status.CREATED).entity(result).build();
     } catch (IllegalArgumentException e) {
       return RestUtil.R.badRequest(e.getMessage());
@@ -2608,5 +2623,12 @@ public abstract class AbstractEventEndpoint {
     }
   }
 
+  private void checkAgentAccess(final String eventId) throws UnauthorizedException, SearchIndexException {
+    final Opt<Event> event = getIndexService().getEvent(eventId, getIndex());
+    if (event.isNone() || !event.get().getEventStatus().contains("SCHEDULE")) {
+      return;
+    }
+    SecurityUtil.checkAgentAccess(getSecurityService(), event.get().getAgentId());
+  }
 
 }
