+++ modules/admin-ui/src/main/java/org/opencastproject/adminui/endpoint/CaptureAgentsEndpoint.java
@@ -35,11 +35,15 @@ import static org.opencastproject.util.doc.rest.RestParameter.Type.STRING;
 import org.opencastproject.adminui.util.TextFilter;
 import org.opencastproject.capture.CaptureParameters;
 import org.opencastproject.capture.admin.api.Agent;
+import org.opencastproject.capture.admin.api.CaptureAgentAdminRoleProvider;
 import org.opencastproject.capture.admin.api.CaptureAgentStateService;
 import org.opencastproject.index.service.resources.list.query.AgentsListQuery;
 import org.opencastproject.index.service.util.RestUtils;
 import org.opencastproject.matterhorn.search.SearchQuery.Order;
 import org.opencastproject.matterhorn.search.SortCriterion;
+import org.opencastproject.security.api.SecurityService;
+import org.opencastproject.security.api.UnauthorizedException;
+import org.opencastproject.security.util.SecurityUtil;
 import org.opencastproject.util.NotFoundException;
 import org.opencastproject.util.SmartIterator;
 import org.opencastproject.util.data.Option;
@@ -94,6 +98,10 @@ public class CaptureAgentsEndpoint {
   /** The capture agent service */
   private CaptureAgentStateService service;
 
+  private CaptureAgentAdminRoleProvider roleProvider;
+
+  private SecurityService securityService;
+
   /**
    * Sets the capture agent service
    *
@@ -102,6 +110,14 @@ public class CaptureAgentsEndpoint {
    */
   public void setCaptureAgentService(CaptureAgentStateService service) {
     this.service = service;
+  }
+
+  public void setRoleProvider(CaptureAgentAdminRoleProvider roleProvider) {
+    this.roleProvider = roleProvider;
+  }
+
+  public void setSecurityService(SecurityService securityService) {
+    this.securityService = securityService;
   }
 
   @GET
@@ -203,11 +219,16 @@ public class CaptureAgentsEndpoint {
   @RestQuery(name = "removeAgent", description = "Remove record of a given capture agent", pathParameters = { @RestParameter(name = "name", description = "The name of a given capture agent", isRequired = true, type = RestParameter.Type.STRING) }, restParameters = {}, reponses = {
           @RestResponse(description = "{agentName} removed", responseCode = HttpServletResponse.SC_OK),
           @RestResponse(description = "The agent {agentname} does not exist", responseCode = HttpServletResponse.SC_NOT_FOUND) }, returnDescription = "")
-  public Response removeAgent(@PathParam("name") String agentName) throws NotFoundException {
+  public Response removeAgent(@PathParam("name") String agentName) throws NotFoundException, UnauthorizedException {
     if (service == null)
       return Response.serverError().status(Response.Status.SERVICE_UNAVAILABLE).build();
 
+    SecurityUtil.checkAgentAccess(securityService, agentName);
+
     service.removeAgent(agentName);
+
+    // Remove the corresponding capture agent roles
+    this.roleProvider.removeRole(agentName);
 
     logger.debug("The agent {} was successfully removed", agentName);
     return Response.status(SC_OK).build();
@@ -301,5 +322,4 @@ public class CaptureAgentsEndpoint {
     }
     return arr(jsonDevices);
   }
-
 }
